groupe-BESTFOREVER/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
openai
google-generativeai
fastapi-cors
tavily-python
